renamed file, implemented controls
