+++ packages/public/renderer-ui/src/components/index.ts
@@ -6,6 +6,6 @@ export * from './expandable-box'
 export * from './feedback'
 export * from './fetch-dimensions'
 export * from './input-exercise-field'
-export * from './interactive-color-rules'
+export * from './interactive-behaviour'
 export * from './textarea'
 export * from './submit-button'
